+++ src/components/RestartableTimer.test.ts
@@ -1,12 +1,18 @@
 import { render, screen } from '@testing-library/svelte';
 import { expect, it } from 'vitest';
-import { currentScreen, gameTime, hasWon } from '$src/stores/stores';
+import {
+	currentScreen,
+	freeplayGameTime,
+	hasWonFreeplay,
+	shouldResetFreeplayTimer,
+} from '$src/stores/stores';
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
 import timezone from 'dayjs/plugin/timezone';
 import { countdownTimeZone } from '$lib/time';
 import { tick } from 'svelte';
-import Timer from '$com/Timer.svelte';
+import RestartableTimer from '$com/RestartableTimer.svelte';
+import { get } from 'svelte/store';
 
 dayjs.extend(utc);
 dayjs.extend(timezone);
@@ -14,6 +20,7 @@ dayjs.tz.setDefault(countdownTimeZone);
 
 beforeEach(() => {
 	vi.useFakeTimers();
+	shouldResetFreeplayTimer.set(false);
 });
 
 afterEach(() => {
@@ -23,10 +30,10 @@ afterEach(() => {
 
 it('Normal time', async () => {
 	currentScreen.set(null);
-	gameTime.set(5000);
-	hasWon.set(false);
+	freeplayGameTime.set(5000);
+	hasWonFreeplay.set(false);
 
-	render(Timer, { interval: 600, message: 'A lot!' });
+	render(RestartableTimer, { interval: 600, message: 'A lot!' });
 
 	const timer = screen.getByText(/[0-9]{2}:[0-9]{2}/);
 
@@ -42,10 +49,10 @@ it('Normal time', async () => {
 
 it('Out of time', async () => {
 	currentScreen.set(null);
-	gameTime.set(1000000);
-	hasWon.set(false);
+	freeplayGameTime.set(1000000);
+	hasWonFreeplay.set(false);
 
-	render(Timer, { interval: 600, message: 'A lot!' });
+	render(RestartableTimer, { interval: 600, message: 'A lot!' });
 
 	vi.advanceTimersByTime(1000);
 
@@ -64,10 +71,10 @@ it('Out of time', async () => {
 
 it('Has won', async () => {
 	currentScreen.set(null);
-	gameTime.set(5000);
-	hasWon.set(true);
+	freeplayGameTime.set(5000);
+	hasWonFreeplay.set(true);
 
-	render(Timer, { interval: 600, message: 'A lot!' });
+	render(RestartableTimer, { interval: 600, message: 'A lot!' });
 
 	const timer = screen.getByText(/[0-9]{2}:[0-9]{2}/);
 
@@ -83,10 +90,10 @@ it('Has won', async () => {
 
 it('Has won after a long time', async () => {
 	currentScreen.set(null);
-	gameTime.set(1000000);
-	hasWon.set(true);
+	freeplayGameTime.set(1000000);
+	hasWonFreeplay.set(true);
 
-	render(Timer, { interval: 600, message: 'A lot!' });
+	render(RestartableTimer, { interval: 600, message: 'A lot!' });
 
 	const message = screen.getByText('A lot!');
 
@@ -101,10 +108,10 @@ it('Has won after a long time', async () => {
 
 it('Dialog behavior', async () => {
 	currentScreen.set('tutorial');
-	gameTime.set(5000);
-	hasWon.set(false);
+	freeplayGameTime.set(5000);
+	hasWonFreeplay.set(false);
 
-	render(Timer, { interval: 600, message: 'A lot!' });
+	render(RestartableTimer, { interval: 600, message: 'A lot!' });
 
 	const timer = screen.getByText(/[0-9]{2}:[0-9]{2}/);
 
@@ -123,4 +130,37 @@ it('Dialog behavior', async () => {
 	await tick();
 
 	expect(timer.innerHTML).toBe('00:10');
+});
+
+it('Restart behavior', async () => {
+	currentScreen.set(null);
+	freeplayGameTime.set(5000);
+	hasWonFreeplay.set(false);
+
+	render(RestartableTimer, { interval: 600, message: 'A lot!' });
+
+	const timer = screen.getByText(/[0-9]{2}:[0-9]{2}/);
+
+	expect(timer).toBeInTheDocument();
+	expect(timer.innerHTML).toBe('00:05');
+
+	vi.advanceTimersByTime(5000);
+
+	await tick();
+
+	expect(timer.innerHTML).toBe('00:10');
+
+	shouldResetFreeplayTimer.set(true);
+
+	await tick();
+
+	expect(get(shouldResetFreeplayTimer)).toBe(false);
+	expect(get(freeplayGameTime)).toBe(0);
+
+	vi.advanceTimersByTime(5000);
+
+	await tick();
+
+	expect(get(freeplayGameTime)).toBe(5000);
+	expect(timer.innerHTML).toBe('00:05');
 });